LabManager/requirements.txt
Django>=3.0.0
django-macaddress
django-import-export==2.7.1
gunicorn>=20.1.0
django-polymorphic>=3
psycopg2-binary==2.9.1
